fix: replace stub to real Order object in creating order method

+++ services/backend/order-service/src/main/java/com/veteroch4k/order/model/OrderCreatedEvent.java
@@ -1,7 +1,7 @@
 package com.veteroch4k.order.model;
 
 public record OrderCreatedEvent(
-    Long orderId,
-    Long productId,
+    Integer orderId,
+    Integer productId,
     Integer productQuantity
 ) {}
